requirements.txt
pip==23.0.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
einops==0.3.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
fire==0.5.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
fvcore==0.1.5.post20221221
gradio==4.21.0
gradio_client==0.12.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.9.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.1.3
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.1
lazy_loader==0.4
librosa==0.10.1
lightning-utilities==0.3.0
lit==17.0.6
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
numba==0.58.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.1.1
open-clip-torch==2.22.0
opencv-python==4.9.0.80
orjson==3.9.15
packaging==24.0
pandas==2.0.0
parameterized==0.9.0
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.1
pooch==1.8.1
portalocker==2.8.2
proglog==0.1.10
protobuf==3.20.3
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==1.8.3
pytorchvideo==0.1.5
pytz==2024.1
PyYAML==6.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.2
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.3.0
timm==0.9.16
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.0
torchaudio==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
transformers==4.25.1
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
wcwidth==0.2.13
websockets==11.0.3
xformers==0.0.19
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0